Change the level of mocking for the wait_for_stack_ready test
This change fixes the tripleoclient tests, which were broken
by an update in heatclient.

Previously it mocked something that was called by heatclient,
this is out of our control so we can't tell when the logic
changes. This patch updates the test to mock at our bounardy
which is the call to heatclient.

Change-Id: I0b959560ee7df194ec0f4a2b8a1605666fdec551
(cherry picked from commit f0292635fd418bd64fe7fe01fa534565a3591c26)