Add optional overcloud deploy roles_data.yaml override
Allows users to optionally pass -r roles_data.yaml to override the
default file in the plan.

Note this is the first step, a similar interface needs to be added
to the overcloud plan create command but this may require rework
of the default plan creation workflow (or some way to re-trigger
the j2 templating after modifying the roles data).

Change-Id: I5c6e22fe9516c4e2dad4495127d923f971c32f5d
Partial-Bug: #1626955