Deprecate a number of CLI args that map directly to Heat params
For a while now the recommended approach for managing parameters has
been with a Heat environment file that is passed into the deploy
command. However, there are a number of CLI args that still map directly
to parameters. This change marks them as deprecated and includes
direction for users in the help message.

They can then be removed in a later release.

Change-Id: I583521c5c72a5377e9ee3047376c0f7bcf5fbdf6