Allow referencing rendered yaml files in resource_registry
For backwards compatibility we need a fallback when an environment
file references a file that's *.j2.yaml in the user --templates dir
but it's rendered via the plan generation and is then downloaded
to the local temporary tree.  In this case we've got to rewrite
the paths in the resource_registry values to point to the new tree
where the rendered templates exist.

Without this logic, some environment files (such as those enabling
swap for all roles) no longer work, so it's a backwards compatibility
workaround.

Change-Id: I560784dc49842113563f617b343702bb045fca76
Closes-Bug: #1625783