Remove heat_stack_user from the deployer input
heat_stack_user role definitely does not have
enough permission to manage a heat stack and can't
be used in tempest, this is an internal role for heat.

We should use the same role for heat_stack_owner as
for a regular user.

This change also adds the role set for the regular users.
This role set also should enable volume backup usage,
so the swiftoperator role will be used.

Wouldn't be bad idea to rename the swiftoperator role
to something like 'user',
but it is out of scope for this change.

See also:
https://bugzilla.redhat.com/show_bug.cgi?id=1314725
https://bugzilla.redhat.com/show_bug.cgi?id=1383369

Change-Id: If487dab0dc1d187ef0bacb9dbeae1186260ec975