Enable creating images with Glance v2
The main difference between the interface for Glance v1 and Glance v2
is the definition of the image's visibility, which in v1 is defined
by a boolean called 'is_public', while in v2 it's a string valled
'visibility'. This enables the usage of both, depending on the
version used by the deployer.

Change-Id: Ife16a6997c60ddc941527ab51cb3fe2ab0b24c5a