Add a flag to allow the plan to be updated without deploying
This is a stop-gap flag to allow just the plan to be updated. It will
cause the 'deploy' command to do everything up to the point of deploying.

Ideally, there needs to be an 'overcloud plan update' command, but there
currently would be far too much code to duplicate between that command
and the deploy command. They need to be refactored to extract the common
bits into a workflow.

Change-Id: I546babfcbd0d5eafcf777e45187dd493fcfdc1a1
Closes-Bug: #1635301
(cherry picked from commit 5c53f2304007546baab19bd992e2f84cf16871f2)