Only show deprecated args for overcloud deploy
The --*-flavor and --*-scale CLI args were deprecated for the "openstack
overcloud deploy" command. This had the unintentional side effect of making
them as deprecated in the "openstack overcloud profiles match" command despite
there being no alternative method.

This change amends that so that the deprecation warnings are only shown for the
"deploy" command.

Change-Id: Iaf60e1e9c99f50d6e63b7da1712a728cd89559a4
Closes-bug: #1629907