Fix nodes count check on stack-update with custom roles
When using custom roles, there is no requirement that you have a role
called Controller, Compute, ObjectStorage, BlockStorage, or CephStorage.
However, this list is hardcoded in tripleoclient and it is assumed there
are corresponding *Count parameters for these roles names then the nodes
count check validation is performed on stack-update.

This patch parses the roles_data.yaml file if one was specified and gets the
default role count parameters from that file.

Change-Id: Ie28e736917fd54dca5dcac8d2defa51a46d28759
Partial-Bug: #1637204