Reset the parameter_defaults between deployments via the CLI
When the CLI makes subsequent deployments it manually updates the plan
in Mistral. We need to do this for backwards compatibility as the deploy
command previously didn't have any plans or history behind it. So if a
user deploys with a specific parameter or template and then they remove
it in the next deploy it shouldn't still exist. In short, the CLI
expects everything to be passed to the deploy command every time.
Therefore, the easiest way to handle this is by manually resetting the
plan.

This has mostly been the case, but we neglected to reset the
parameter_defaults. Meaning, that a user may deploy with a parameter
like "ComputeScale: 5" and then deploy again, without defining it and it
would use the previous value when they would expect it to use the
default.

Change-Id: I09048a550599146ef9bd81a309d80b10497df7c8
Closes-Bug: #1639787