Default the stack name to 'overcloud' in overcloud node delete
The stack needs to be passed as either a CLI arg with --stack or via the
environment variable OVERCLOUD_STACK_NAME. At the moment if it isn't
provided either way the workflow is called with an empty string for the
stack name. This will then fail.

This change defaults the stack name to 'overcloud' which matches the
deploy command. One of the unit tests was updated so that it didn't pass
the default value to ensure that it can still be overwritten.

Change-Id: Ie44d5cba90e81ae1a35a4cabc6e20a202bfddef0
Closes-Bug: #1640244
Partial-Bug: #1640770
(cherry picked from commit 0af664656d82ab5d658228cd078691f9abc3e789)