Add a new command to get the overcloudrc files
If a user deploys via the GUI, then this file wont automatically be
created. We can now create it via a Mistral action - this command
provides CLI users a way to call that action and save the file.

While adding tests a typo was spotted in utils.py, this is also resolved
by this patch and covered by the tests.

Change-Id: I7808cd2c43a95bbb154ec6c6f77896b97551e354