Clean up image exceptions
When tripleoclient queries for the deployment kernel and ramdisk in
glance, it logs an exception to the console if they don't exist. These
errors are actually redundant since an error is logged later. Also, it
is needlessly verbose on the console, especially when using the
deployed-server.yaml environment where these iamges do not actually have
to exist.

This patch changes the behavior to ignore the exception so that the
traceback is not shown.

Change-Id: I53805d2c805705c8f16e9a6f2ef5df3195f4f70e