Remove python-networking-cisco from image build
The networking-cisco library is currently broken on master due to
deprecated code being removed from neutron. This has already been
removed from tripleo-common[0]. This change should be reverted after
the networking-cisco module is fixed.

[0] https://review.openstack.org/#/c/397959

It should be noted that currently this breaks users using
'openstack overcloud image build'. oooq uses an alternate method for
image building which is why it is not detected in CI.

Change-Id: Ib15fc6b09bc6c703d0e074a5139b9ea09fd27971
Related-Bug: #1631311