Make overcloud image upload honor --image-path
The openstack overcloud image upload command supports --image-path argument
but the code assumes that the command is run from ~/images directory and fails
otherwise with errors such a "The given file overcloud-full.vmlinuz is not a
regular file".

When the code checks if the image file exists we do not see this error as
"self._check_file_exists(os.path.join(parsed_args.image_path,
image))" uses the image_path from user passed arguments. However, when
building the file name for uploading the image, the --image-path is
neglected.

This commit adds code to make the image upload command work from
any directory as long as --image-path option is passed.

Closes-Bug: #1647519

Change-Id: I74a21d73da95e81c5270f7264c25eaee923ba9ba
Co-Authored-By: Ben Nemec <bnemec@redhat.com>