Fix plan delete
As part of the overcloud delete implementation, we refactored the plan
delete to be a shared function. Unforunately the wrong inputs were being
passed to the workflow so it would return an Invalid input error. This
change fixes the input so that the container is properly passed to the
workflow as an input.

Change-Id: I33032c9e6f4a8cf002e65ebab60752c44fb5fdb2
Closes-Bug: #1647747