Simplify Fetching the passwords from Mistral
This patch removes some of the indirection, deception and complexity in
tripleoclients password handling. At the end of the Newton cycle the
password generation was moved to Mistral. tripleoclient was updated, but
some methods got left behind in the process.

Before this patch, we had one utils function, that called another, that
called a workflow wrapper. The tests then embraced this oportunity and
mocked at different levels of abstraction, some of the tests even mocked
it out in mutliple places to be safe.

This patch replces that with a small function in overcloud_deploy. Once
the post-deploy steps are removed, there will be less of a need for the
password generation and it can hopefully be deprecated with that code.

Change-Id: Icf77dae225caa716c35ce5ded83f7356c958d695