Add new hiera agent hook to legacy image build
The change Ia1864933235152b7e899c4442534879f8e22240d added these to the
newer overcloud-images.yaml method of building images.  Unfortunately
because the old 'openstack overcloud image build --all' method does not
leverage this file yet, it leads to a timeout in deployment because the
heat agents are not available.

Change-Id: I00b95e9fa890312b4bfda7f097d2e44bed0b22ca
Closes-Bug: #1651616