Provide an option to disable password generation on deployment
Operators may choose to use the passwords provided by them and
does not intend to generate during the deployment. Provided an
option to disable password generation if requested by user explicity.
It will be helpful to identify if there are any new passwords added
which has to be added during the deployment update by the operator.
By default, the password generation will be enabled.

Partial-Bug: #1611704
Depends-On: I141a0727db9f19f80712e2c40069622ecc22d969
Change-Id: Iedfd7fd1456ed61cb9208532d15bdb63fbc01de9