Switch undercloud upgrade to use upgrade script
In order to better handle upgrades, we need to be able to know when
we are performing an upgrade instead of an installation. This change
switches the undercloud upgrade process to use the
instack-upgrade-undercloud script which will expose the fact that we are
running an upgrade process to the underlying puppet scripts.

Change-Id: I493204117ca158e99f706770a02fbb7f6392673b
Depends-On: Ie3cb21e30334fe8ffc0a9d6e707b42269b64c9ec
Related-Bug: #1649341
Related-Blueprint: undercloud-upgrade