Fix mitaka unit tests
There are a couple of issues here.  First, we have an uncapped
openstackclient dependency, which means we are now pulling in a
too-new version of osc without the utils module.  The solution to
this is to start using upper-constraints, but that brings us to
problem two.

Problem two is that we had some requirements that pointed directly
at git repos in mitaka, which is incompatible with upper-constraints.
To fix this, we can now use the released versions of these deps.
os-cloud-config is in global-requirements without a version
specified, and tripleo-common is added as a dependency of this
patch.

Change-Id: I1c219a3547ca48e5c4a649441632eabc326081f9
Closes-Bug: 1656871