Call undercloud install function directly
For about 2 years we've been indirectly calling the undercloud
install function through a bash script.  Originally this was done
to allow for some backwards compatibility with mismatched packages,
but later it was kept because a direct call broke logging.  This
is no longer a problem, probably due to some changes in how logging
in instack-undercloud works.

Change-Id: If07880e4bc25cffe485704772d64f62288a1a53d
Depends-On: Iff76f639bcacb5bcc1274fd837e47d34041afa15
Related bp undercloud-upgrade