Mock os.path.isfile correctly
Previously we were monkey-patching os.path.isfile with a mock in a
way that was not undone at the end of the test.  This meant the
mock leaked into other tests and could cause unpredictable
behavior.  The correct way to handle this is to use mock.patch,
which will be automatically removed at the end of the test.

Change-Id: Ie86f817f37fe40675934a3e9a05ea0ab948dddf1