Show stack failures on error
On stack create or update error, go ahead and run "openstack stack
failures list overcloud". It's the best way to show the immediate error
from a stack failure and it's more user friendly to go ahead and show
the error.

Note that this module is not currently tested with any unit tests, but
if there is general agreement on this patch, I'll work on adding some in
a followup.

Change-Id: Ibe9a97f3a3088954b1e88fa7c741845f70bd6edf