Remove calculation for NeutronDhcpAgentsPerNetwork
This patch removes the calculation that set the
NeutronDhcpAgentsPerNetwork configuration based on the number of
"controllers" being deployed. With composable roles, this is incorrect
and the default, if required, should be determined by the number of
neutron dhcp agent services being deployed - not a role count.

Closes-Bug: #1632721
Depends-On: I06628764c4769d91bbc42efe1c722702d6574d02
Change-Id: If202a8255f8ad9f662f0d66a5f630f1b7ae60a26
(cherry picked from commit 1051880e92a46a508bc92da01f4beb7a9203525b)