Ignore os-cloud-config deprecation warning
tripleoclient users really have no control over whether the code imports
os-cloud-config or not and shows a deprecation warning. We still have to
rely on os-cloud-config in case someone is deploying old templates. But
in the meantime, we should not spam the user with the deprecation
warning every time they use tripleoclient.

Change-Id: I54e4c6f117c587fee05fcd5fc3f5c9feee61785f
Closes-Bug: #1670817