Use tmp env path when rewriting absolute resource paths
When rewriting absolute resource paths from a resource_registry that
failed to load due to missing resources (one case being where those
resources are rendered j2), then we should refer to the initial
environemnt that failed to load by the tmp path.

This is required since itself may have been rendered from j2. In such a
case, the environment doesn't exist under the absolute template root
path as it's not been rendered to that location.

Change-Id: Ib681729cc2728ca4b0486c14166b6b702edfcaab
Closes-Bug: #1671096