Mark 'undercloud deploy' as experimental
This patch marks the 'undercloud deploy' feature as experimental
for Pike. There is not yet agreement on when this will (if ever)
replace instack-undercloud. Until then I think marking it as
experimental is wise to avoid confusion should we release Pike
without converting 'undercloud install' to make use of this.

Change-Id: Id6f1343994429db2feb8250332dd9e1c62b73fde