New command "overcloud container image upload"
This is the container image equivalent to the "overcloud image upload"
command. It wraps tripleo-common logic which has existed for some time
now[1].

The configuration file used in CI is also in tripleo-common [2] and
the existence of this command will allow the custom python[3] to be
replaced with a simple command invocation.

The same configuration file format will be used to implement another
new command "overcloud container image build".

[1] http://git.openstack.org/cgit/openstack/tripleo-common/tree/tripleo_common/image/image_uploader.py
[2] http://git.openstack.org/cgit/openstack/tripleo-common/tree/contrib/overcloud_containers.yaml
[3] http://git.openstack.org/cgit/openstack/tripleo-quickstart-extras/tree/roles/overcloud-prep-containers/templates/upload_images_to_local_registry.py.j2

Change-Id: Id505b57698b6a15c62494764b2d9eca1933e63f1