create history for the overcloud deploy command
Add a history directory in user $HOME for tracking
all the calls of the openstack overcloud deploy command
passed by the users.
Each deploy command run by a user is now track in an history file under
$HOME/.tripleo directory.
It allow to retrieve all the previous calls from
the openstack overcloud deploy command, the environment files used and the
templates directories used, for analysis, debugging or tracking. See
https://bugs.launchpad.net/tripleo/+bug/1673700

It will be very usefull for post deployment user experience and
for customer debuging.

This review aim to be backport for old releases support.
I will implement a  mistral workflow in a follow up review for
Pike, as suggested by D0ugal (D Matthews)

Change-Id: Ibf7dd4a7b0036b6f2a0809d0448a05665c809167
Closes-Bug: #1673700