Move clients into class constructor
Currently we reference the various clients in several functions, so it's
cleaner if we just define them in the constructor vs passing them around
between functions etc.  In particular the swift client will be needed
in several functions if we're to break up the plan-update logic related
to handling environments effectively, so we may as well move all clients
for consistency (when more validations are done via mistral we can probably
remove some of these altogether e.g ref the validations-in-workflows
blueprint).

Change-Id: I8e3ca726ec6d873529127c2326f1b65dbb2bc22e
Partial-Bug: #1635409