Write user parameters environment to swift
Currently we only merge this into the local merged environment,
but instead we should write it as a separate user-environment in
swift, which will enable heat server side parameter merging to
work in future.  The _write_user_environment function is added to
handle writing the file to the local tmpdir tht_root, and uploading
the same data to the plan.

Partial-Bug: #1635409
Change-Id: I74bde48ef18c2e59233b7535578f394fbd1f7553