Avoid race condition when setting up websocket
There is a potential for a race condition if the websocket is not set
up before the call is made to Mistral. The websocket setup can take
sufficient time to set up, and if an error is thrown immediately from
a workflow, it has the potential of returning a message, and having
the websocket consume it before the websocket client is even set up.

Also, I think this has the benefit of keeping the call to Mistral
from even being made if there is something wrong with the websocket
setup.

Change-Id: Ib331037a7f5f4e59862d2b9646a83acdb18313eb