Add --skip-deploy-identifier
Add a new cli argument, --skip-deploy-identifier. The argument will
eventually be passed to the tripleo-common DeployStackAction action to
disable setting a unique value for the DeployIdentifier parameter, which
means the SoftwareDeployment resources in the templates will only be
triggered if there is an actual change to their configuration. This
argument can be used to avoid always applying configuration, such as
during node scale out.

Removes the setting of DeployIdentifier completely from tripleoclient,
as that is now handled by the tripleo-common action.

Depends-On: Idb901a841846fec33d189b3c95f669b0380498ce
Change-Id: I294eabe84e070367981534030b5927062f00c628