Provide commands for example roles
In order to improve the experiance for the end user, we have existing
sets of roles that can be combined in a roles_data.yaml file for the
deployment. This change exposes the ability to generate roles_data.yaml
from a folder containing role yaml files rather than requiring the end
user manually construct the roles_data.yaml.

Change-Id: I326bae5bdee088e03aa89128d253612ef89e5c0c
Related-Blueprint: example-custom-role-environments