Add list options to overcloud container image build
This change adds the options --list-images and --list-dependencies to
the "overcloud container image build" command.

With these options, images will not be built, but the command will
return structured data for what images would be built.

This provides parsable data to allow other standard image building
pipelines to do the actual image building.

The order of --list-images is determined by doing a depth-first
traversal of the dependencies structure, which provides a reasonable
order for image building.

Change-Id: Ibabbb487465566dc2550a9b89e76800dfcae8257