Add check for disable_upgrade_deployment flag in roles_data
Check that the disable_upgrade_deployment flag is set in any
operator provided roles_data.yaml file (the --roles-file
parameter).

This is an alternative proposal to
I9327f42bfbebf050f79224ee1c1ff41e49b9c2fd and the advantage
here is that we can use the existing validation_warnings_fatal
to raise the exception or stay at logging a warning. The
validation_warnings_fatal defaults to False.

Related-Bug: 1699792
Change-Id: I606a3010e7ecd83f34b852a35ed298c30f65adeb