Don't filter config keys by default
Having a hard-coded list of keys in the client is fragile and also
makes it non-obvious when you add a new key to t-h-t and nothing
appears when dumping the config via tripleoclient.

Instead lets dump all keys by default, and offer the option of
filtering by key if needed.

Change-Id: I53cacf7a085fe6699539cc08ead5c12d392cdc33