Removing the default role to swiftoperator
Not always the swiftoperator role is created, only when you have
the swift proxy template enabled. One of the examples is the
scenario001 that don't have swift proxy installed. On that cases
the role set in tempest-deployer-input became invalid.
There's no need to set the operator_role and tempest_roles if you
set the use_dynamic_credentials to true by default. In this case
tempest will handle the creation of the needed roles for each test
that requires a role to be created. This will fix the cases where
we have the swiftoperator role set, but not created.

Change-Id: Ib114241629dfd75bb1277514e3ca5730e9202314