Check for stack failure earlier in undercloud deploy
For the undercloud deploy, if the stack fails before the deployed-server
resource is created, it will take 30 minutes to time out since there is
no check for the stack status. Further, the exception is not as helpful
as it could be about what failed as not all the events were shown.

This patch adds a check within the loop looking for the deployed-server
resource. It will now fail immediately if the stack fails before the
deployed-server resource is created. It also again outputs the stack
events. Some may be redundant, but it will at least show the errors now.

Change-Id: I9fa3239d688ce83676ecfab2f3d102f9e48a43a5