Adds when in upgrade_tasks playbook written by config download
As part of the bug below, this can be used to make sure the
playbooks generated from the heat output can be iterated over
using the loop variable.

This adds a pre-processor for upgrade_tasks that adds a "when
step == N" condition based on the value of the tags. That is
"tags: step1" becomes "when: step == 1". When there is an
existing when statement the new step condition is appended.

Change-Id: Ief593dc758a2ffe33c1cbcbda9289393fcf023e4
Related-Bug: 1708115