Filter container images by deployed services
This change adds the --service-environment-file/-e argument to the
prepare command which will take heat environment files that specify
which services are containerised.

When specified, the resulting image entries will be filtered only by
the containerised services discovered in the environment.

-e is used as the short argument so that tools like quickstart can use
the same variable to specify these environments for both the prepare
command and the overcloud deploy command.

Once this change merges, any calls to prepare in CI can be modified to
reduce the number of images being transferred for 'upload' calls. For
example, specifying -e environment/docker.yaml will reduce image
transfers from 85 to 44.

Change-Id: I35866fb24915ef80b86cc402015afb08cb155dcf
Partial-Bug: #1710992
Depends-On: #Ie8801b11921c46923b0f7c9aaba6bf524c464e82