Command to discover the versioned tag from latest
This change implements the command
"openstack overcloud container image tag discover" which will become
part of the update and upgrade workflow for deploying new container
images.

It is used to discover the version-based tag by inspecting the image
from a stable tag like current-tripleo-rdo. Stable tags like 'latest'
or 'pike' can't be used for container updates because something needs
to change to trigger the new containers being pulled. Without this
command it would be up to the user to find out what versioned tag to
specify when calling prepare.

This was implemented as its own command instead of integrating with
prepare because there may be multiple image build chains feeding into
the image list (such as RDO, ceph and eventually others) and each will
need its own call to discover the versioned tag.

  tag=$(openstack overcloud container image tag discover \
        --image trunk.registry.rdoproject.org/pike/centos-binary-base:current-tripleo-rdo
        --tag-from-label rdo_version)
  echo $tag
  openstack overcloud container image prepare --tag $tag

Change-Id: I12b16cb267c80e3059786fb980178eb5b3d1a76d
Depends-On: I27ea031287604d70032fb5392aecbce313d4b096
Closes-Bug: #1708967