Add neutron_driver value to prepare command
OpenDaylight/OVN is not enabled by default in overcloud. A logic needs to be
implemented to create ODL image reference when ODL needs to be deployed and
the same for OVN. This patch adds neutron driver value set depending on the
environment file specified to image prepare command.

Co-Authored-By: Numan Siddique <nusiddiq@redhat.com>
Change-Id: I3e395986ad777b1d6b057bb33fac8d5ed2d3d26a
Partial-Bug: #1713685
Partial-bug: #1699085