Convert step to integer in when statement for upgrade tasks
Currently the when conditionals in the upgrade tasks aren't evaluated
correctly and as a result the upgrade tasks are skipped. This change
converts the step variable in the when statement to an integer to get
it evaluated properly.

Related-Bug: 1708115
Change-Id: I4ee1a2729d74442570f1b1f38b0d03a95ea7793f
(cherry picked from commit 53f79082ca89937cf5997e62f472f436cf886884)