Filter out disabled services from prepare command
Previously, the `openstack overcloud container image prepare` looked
only at the resource registry -- checking if a resource pointed to
a containerized service -- to determine which images to use without
looking at the list of services that are enabled or not per roles.

This commit introduces a new '--roles-file' option to the  command to
specify the path to the roles file when not using the default.

Closes-Bug: #1714296
Change-Id: I0188df062bd341dcf4791bfa9c9d0f1ad7f9396e