Autodetect insecure registry in container image prepare
This change automatically inserts the DockerInsecureRegistryAddress
parameter into the containers image environment file if the registry
is reachable and unencrypted.

This eliminates a documented manual step, which is also required
throughout CI. It has been done in tripleoclient so that it can be
backported. The dedicated workflow for
blueprint container-prepare-workflow will have its own insecure
detection.

Change-Id: I74586145c222921880402cec92bbaa5f10e8b972
Closes-Bug: #1722632