Add --output-dir
Add an --output-dir cli arg to be used alongside --config-download. The
arg controls where the downloaded ansible playbooks and log files will
be stored on the local filesystem.

The path must be writeable by the current user and mistral user, so it
will typically need to be pre-created manually with the correct
permissions.

Depends-On: I89b17d0c4f035705fd5230caad06d331fd5b5e76
implements: blueprint ansible-config-download

Change-Id: Iff7a1b211a545f1227b9e13e304b858128fb855b