Ignore missing rendered heat envs in prepare
Since we use j2 templates for some of the heat environments, and at the
moment they are rendered when uploading them to swift at the beginning
of the deployment workflow, it is possible that one of the heat
environment file passed to the prepare command is, or references a file
that is, not yet rendered.

This commit implements a custom object_request() function that makes
process_multiple_environments_and_files() ignore the missing files.

Change-Id: I6f31898ca3db9563d3d8e8e6e0a9f0cdaebf24ca
Co-Authored-By: Thomas Hervé <therve@redhat.com>
Closes-Bug: #1730650
(cherry picked from commit a68cbfd5157ec367020a753905f7e1802d127a0f)