Set also up-to-date overcloud flavor parameters
Previously we've been setting just the deprecated variants of the
flavor parameters, which could cause problem with custom
role_data.yaml if it didn't include the appropriate
deprecated_param_flavor key for Controller or ObjectStorage roles. Now
we set both the deprecated (for old version support) and the
up-to-date parameter variant (for custom role data).

Change-Id: I53f1a7da3edf4accb9552bb7125e99e468124856
Closes-Bug: #1733327