Add entry per enabled service in the inventory
Expand the static inventory with an entry for each of the enabled
services. This allows to access the host/address for the services just
like it would be done in the overcloud.

An example of a scenario where this is needed is deploying services on a
kubernetes that has been deployed in the undercloud with tripleo. To
know where kubernetes is, we need to be able to get the hostname/address
for `kubernetes_master`

Change-Id: Ia32cef7da5187d3fe61bc168dd753e3b609e6938