Add major update implementation
Add the openstack overcloud upgrade command and workflow
for major upgrade from P to Queens.
It share the same code base as the minor upgrade but add options
for the converge part and take as parameter a list of
environment files, that could be needed in some cases for the
upgrade init stage. (like repo_setupt)

Change-Id: I54f8fc57b758e34c620d607be15d2291d545ff6f