undercloud_deploy: add opts to setup virtual-ips
This patch wires in --control-virtual-ip and --public-virtual-ip
options that can be used to configure VIPs during the undercloud
install. Use of these options would need to be coordinated with
specific heat environments that enable the parameters we set for
this to do their job. See inline comments in this patch for
details on these environments.

Co-Authored-By: Ian Main <imain@redhat.com>

Change-Id: Ic226fdef1f5d09fbefc9fd878ed54fb82d905aef