Add --validate-only to openstack overcloud node import
Previously there was 'openstack baremetal instackenv validate' to
quickly validate an environment file.

Much of which is now done in the validate_nodes mistral workflow.  Add a
mode to cease operations after validating the environment file

Note on the backport:
The old command was removed in Queens, so we need to backport this
new flag to fill in the gap. The workflow was introduced in Pike.

Closes-Bug: #1768604
Change-Id: Ie2bfc79b45f4652285718be34a5e07ce5777c60d
(cherry picked from commit ebc64b681cbd0e760d34b1dcc268f3b487a051e8)