Handle user-provided TLS certificate/key for the undercloud
instack-undercloud has support for using a user-provided certificate and
key. This is done through the undercloud_service_certificate
configuration option. This commit adds support for the same
functionality for the containerized undercloud.

It relies on pyca/cryptography, so I added it to the requirements.txt
file. And what it does is that it will read the PEM file and add the
parameter_defaults that are needed to enable TLS.

Change-Id: I26dbde57131f6328d0c5e400dcdfdf62cdd43a41