Add --platform support when uploading images
Adding --platform when uploading images will set a tripleo_platform
image meta data property and prefix all images with the platform (and
arch) to make them stand out in an openstack image list.

Support is added in such a way that that image names are not altered if
no --platform is specified.  As platform really only makes sense
*within* and architecture we ensure that both options are supplied.

Blueprint: multiarch-support
Change-Id: Ia56d22b244cfa5dddb5d110b54a525dfd80a7aaf