Composable role and service fails workflow
When defining a resource_registry for a composable service, the workflow fails
because it validates before the user-environment is added.
This patch moves the workflow invocation after the creation of the user-
environment.
This is caused by https://review.openstack.org/#/c/511249/ which changes the
behaviour of the action in tripleo-common.

Change-Id: Idf17e3f817708d9e7dd7241063434ce56b6be052
Closes-Bug: 1743575