Switch to using assert_has_calls() instead of assertEqual()
While adding multiarch support I noticed that some tests I was copying
had switched the values of actual and reference.  While this doesn't
alter the correctness of the tests it is somewhat strange when the tests
fail and a human is looking at the output.

Also during review Steve Baker noted we could use assert_has_calls() so
avoid confusing humans by switching :)

Change-Id: I996248f06ff47a7e20398848ae047eb4c2227dec