undercloud: introduce deployment_user parameter
deployment_user is a new parameter, default to getpass.getuser(),
will feed DeploymentUser parameter in THT, primarly used to add the
user to the 'docker' group, so our operators can run the `overcloud
container` commands when the undercloud is containerized.

Right now, the docker group was modified in puppet with
instack-undercloud but with the heat-installer we want to do it in
tripleoclient. DeploymentUser was chosen to be "standard" and can be
used for both undercloud & overclouds notions.

Change-Id: Ia5cc7b34ebee8cf2f49300ce23050370d5f1038a