Populate ContainerImagePrepare variable
If the user does not specify an explicit container_images_file in the
undercloud.conf, the variable ContainerImagePrepare will be set with
the values required to run the image prepare during the undercloud
deploy run.

This uses the same ContainerImagePrepare variable interface proposed
for the overcloud mistral-driven container prepare
I1b485ff9491fe2b1cf850116f9133de8d05f3d40.

The actual image preparation will be added to undercloud_deploy.py in
a following change in this series.

Change-Id: Ie8543bed3363aac539cfd35e04e8dfaaa166ebc3
Blueprint: #containerized-undercloud