Do container image prepare during undercloud deploy
This change consumes the ContainerImagePrepare parameter during
"openstack undercloud deploy" to populate the container image
parameters. This includes doing tag discovery if tag_from_label is
set, and image uploads if push_destination is set.

Rather than creating a separate environment file for these parameters,
the parameters are set directly in the environment if that parameter is
not already set.

Change-Id: I00f72b1cd7ab6a03dd42bc489f3c950ee32dedeb
Blueprint: containerized-undercloud