uc/deploy: try to keep passwords from undercloud-passwords.conf
When upgrading a non-containerized undercloud to a containerized
undercloud, we would like to keep the same passwords that were used
before so our operators don't have to deal with new passwords and the
upgrade is more transparent.

This patch is doing the following:

If undercloud-passwords.conf exists, undercloud-deploy will generate
Heat parameters based on the existing passwords and override what was
generated by the password workflow run by tripleo-common.

Closes-Bug: #1753063
Change-Id: Ib2d944c6ac7433461575ca6e9578e39404258b7b
Co-Authored-By: Thomas Herve <therve@redhat.com>