Make container_images_file required
When doing a containerized undercloud deployment, it is required to
pass in a file containing the containers to use. Previously there was
a default value meaning if you didn't pass in the file it would start
but fail during the deployment because the containers were not defined.
This parameter is required until we come up with a solution to
automatically generate this file during the deployment.

Change-Id: I2a5694f3dd623672c2841ccfdcf1f6325509d806
Partial-Bug: #1754067