Wire up a --skip-tags parameter for overcloud upgrade run cli
This exposes the --skip-tags which is useful for skipping the
service validations (--skip-tags validation). The expected
format is a comma-separated string for multiple values,
though 'validation' and 'pre-upgrade' are the only currently
supported tags for --skip-tags in the upgrade playbooks. The
full list of tags currently used in tripleo-heat-templates
upgrade_tasks is at [0]

  openstack overcloud upgrade run --nodes foo
                                  --skip-tags validation

[0]: https://github.com/openstack/tripleo-heat-templates/blob/3eb0c62e47fd1635900913b7f0ebbe134b77c7ad/tools/yaml-validate.py#L167 
Change-Id: Ie7fb8d9a388c6d53a31800406b03ddb8ed426401
Depends-On: I8544de64d3307e3dc925c1cecf2d9156e31d25a8